+++ golang/build.rs
@@ -5,11 +5,7 @@ use std::env;
 fn main() {
     let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
 
-    cbindgen::Builder::new()
-      .with_language(cbindgen::Language::C)
-      .with_header("/*dellemc license*/")
-      .with_crate(crate_dir)
-      .generate()
-      .expect("Unable to generate bindings")
-      .write_to_file("./test/pravega_client.h");
+    cbindgen::generate(crate_dir)
+        .expect("Unable to generate bindings")
+        .write_to_file("./pkg/pravega_client.h");
 }
