--- golang/client_go/go.mod
@@ -1,4 +1,0 @@
-module dellemc/pravega/goclient
-
-go 1.17
-
