+++ golang/go.mod
@@ -1,0 +1,3 @@
+module pravega-client
+
+go 1.17
